Change structure of files to have go19 and go110 variants for conn, driver and stmt as needed
