+++ conn_go19.go
@@ -6,18 +6,9 @@ import "database/sql/driver"
 
 var (
 	_ driver.NamedValueChecker = wrappedConn{}
-	_ driver.NamedValueChecker = wrappedStmt{}
 )
 
 func (c wrappedConn) CheckNamedValue(v *driver.NamedValue) error {
-	if checker, ok := c.parent.(driver.NamedValueChecker); ok {
-		return checker.CheckNamedValue(v)
-	}
-
-	return driver.ErrSkip
-}
-
-func (c wrappedStmt) CheckNamedValue(v *driver.NamedValue) error {
 	if checker, ok := c.parent.(driver.NamedValueChecker); ok {
 		return checker.CheckNamedValue(v)
 	}
