+++ driver_go110.go
@@ -1,0 +1,23 @@
+// +build go1.10
+
+package instrumentedsql
+
+import "database/sql/driver"
+
+var _ driver.DriverContext = wrappedDriver{}
+
+func (d wrappedDriver) OpenConnector(name string) (driver.Connector, error) {
+	driver, ok := d.parent.(driver.DriverContext)
+	if !ok {
+		return wrappedConnector{
+			parent:    dsnConnector{dsn: name, driver: &d},
+			driverRef: &d,
+		}, nil
+	}
+	conn, err := driver.OpenConnector(name)
+	if err != nil {
+		return nil, err
+	}
+
+	return wrappedConnector{parent: conn, driverRef: &d}, nil
+}
